requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Columnar==1.4.1
deep-translator==1.11.4
frozenlist==1.4.1
futures==3.0.5
goslate==1.5.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.7.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
libretranslatepy==2.1.4
lxml==5.2.2
multidict==6.0.5
py-cord==2.6.0
Py-Dictionary==4.1.4
PyDictionary==2.0.1
python-dotenv==1.0.1
requests==2.32.3
rfc3986==2.0.0
setuptools==70.3.0
sniffio==1.3.1
soupsieve==2.5
string-color==1.2.3
toolz==0.12.1
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
